Move Stripe metadata error predicate to adapter shared
